refactor(services): restore projection failure logging, shrink evaluations index
Review fix round 1 for Task 14.

_mutation_projection.py: _project_all_runs deferred-imports
mutation_rescore._logger at the failure site when no log is injected,
restoring the exc-path warning that NULL_LOG had silently swallowed
in production (the call site can't pass log= directly -- two tests
patch the whole function with a bare single-arg side_effect).

_evaluations_index.py: shrunk from 373 to 280 lines. Extracted three
self-contained free functions (none patched by any test) into a new
_run_index_fs.py sibling: _remove_run_directory (delete()'s rmtree
blocks), _scan_reports_root_for_run (get_log_run_dir()'s filesystem
fallback), _external_job_is_complete (is_complete()'s ext- branch).
_merge_internal_jobs also moved there for the same reason. _logger is
threaded into _remove_run_directory (same-module reference, no import
cycle). _status_json_terminal moved to _run_status_readers.py (both
_evaluations_index.py and the new _run_index_fs.py need it).

Baseline: 152 -> 151 (removes _evaluations_index.py:1:file).
